+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: mraymond <mraymond@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/29 18:27:32 by gvial             #+#    #+#             */
-/*   Updated: 2022/10/07 16:52:47 by mraymond         ###   ########.fr       */
+/*   Updated: 2022/10/07 18:43:55 by mraymond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -85,10 +85,10 @@ typedef struct s_cmd
 {
 	char			*cmd_path;
 	char			**args;
-	char			*fd_in;
+	char			**fd_in;
 	char			**fd_out;
 	int				append;
-	int				heredoc;
+	int				*heredoc;
 	struct s_cmd	*next;
 }	t_cmd;
 
@@ -104,6 +104,7 @@ typedef struct s_ms
 	int		cmd_index;
 	int		nb_cmd;
 	int		*child_id;
+	int		err_num;
 }	t_ms;
 
 //==============================================================================
@@ -137,12 +138,35 @@ char	**add_env_var(char **envp, char *var_name);
 
 //05_exec.c
 void	exec(t_ms *ms);
+void	fd_allocation(t_ms *ms);
+void	child_creation(t_ms *ms);
+void	waiting_n_closefd(t_ms *ms);
+int		child_process_to_index(t_ms *ms, int waitpid_return);
+
+//05_child_exec.c
+void	child_execution(t_ms *ms);
+t_cmd	*cmd_lst_index(t_ms *ms, int cmd_index);
+void	pipe_redirection(t_ms *ms);
+void	exec_fail(t_ms *ms);
+void	close_keep_errno(int fd);
+
+//05_redirection.c
+int		redirection_in(t_cmd *cmd);
+int		here_doc(char *str_eof);
+int		open_fd_in(char *filename);
+int		print_open_err(char *filename, int error);
+void	redirection_out(t_cmd *cmd);
 
 //01_valid_line
 int		valid_line(char *line);
 int		print_line_err(int error);
 
 // utils
+int		lst_len(t_cmd *head);
+
+//utils2
+void	free_dbl_ptr(void **ptr, int option);
+void	free_ms(t_ms *ms, int with_lstcmds);
 
 //==============================================================================
