+++ src/03_utils.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   lst_utils.c                                        :+:      :+:    :+:   */
+/*   03_utils.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gvial <marvin@42quebec.com>                +#+  +:+       +#+        */
+/*   By: mraymond <mraymond@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/29 18:28:28 by gvial             #+#    #+#             */
-/*   Updated: 2022/09/29 18:28:29 by gvial            ###   ########.fr       */
+/*   Updated: 2022/10/07 17:36:57 by mraymond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
