+++ src/05_exec.c
@@ -6,7 +6,7 @@
 /*   By: mraymond <mraymond@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/29 18:28:01 by gvial             #+#    #+#             */
-/*   Updated: 2022/10/07 16:57:24 by mraymond         ###   ########.fr       */
+/*   Updated: 2022/10/07 18:39:12 by mraymond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,28 @@
 
 void	exec(t_ms *ms)
 {
-	ms->nb_cmd = ft_lstsize(ms->cmds);
+	ms->nb_cmd = lst_len(ms->cmds);
 	ms->child_id = (int *)ft_calloc(ms->nb_cmd, sizeof(int));
 	fd_allocation(ms);
 	child_creation(ms);
 	waiting_n_closefd(ms);
+}
+
+void	fd_allocation(t_ms *ms)
+{
+	int	i;
+	int	temp_pipe[2];
+
+	ms->pipe = (int *)malloc(sizeof(int) * ((ms->nb_cmd - 1) * 2 + 2));
+	ms->pipe[0] = 0;
+	ms->pipe[(ms->nb_cmd - 1) * 2 + 1] = 1;
+	i = -1;
+	while (++i < ms->nb_cmd)
+	{
+		pipe(temp_pipe);
+		ms->pipe[i * 2 + 1] = temp_pipe[1];
+		ms->pipe[i * 2 + 2] = temp_pipe[0];
+	}
 }
 
 void	child_creation(t_ms *ms)
@@ -35,18 +52,6 @@ void	child_creation(t_ms *ms)
 	}
 	if (process_id == 0)
 		child_execution(ms);
-}
-
-int	child_process_to_index(t_ms *ms, int waitpid_return)
-{
-	int	i;
-
-	i = 0;
-	while (ms->child_id[i] && ms->child_id[i] != waitpid_return)
-		i++;
-	if (!ms->child_id[i])
-		return (-1);
-	return (i);
 }
 
 /* main process wait for each child to finish
@@ -70,151 +75,22 @@ void	waiting_n_closefd(t_ms *ms)
 			{
 				close_keep_errno(ms->pipe[(child_index * 2) - 1]);
 				close_keep_errno(ms->pipe[(child_index * 2)]);
-				if (WEXITSTATUS(status) != 0 && child_index < ms->nb_cmd - 1)
-					close_keep_errno(ms->pipe[(child_index * 2) + 1]);
 			}
+			if (child_index == ms->nb_cmd - 1)
+				ms->err_num = 0;
 		}
-		else
-			printf("ERROR EXITING CHILD PROCESS, NOT NORMAL\n");
+		//read in pipe_err
 	}
 }
 
-void	close_keep_errno(int fd)
-{
-	int	temp_errno;
-
-	temp_errno = errno;
-	close(fd);
-	errno = temp_errno;
-}
-
-void	fd_allocation(t_ms *ms)
+int	child_process_to_index(t_ms *ms, int waitpid_return)
 {
 	int	i;
-	int	temp_pipe[2];
 
-	ms->pipe = (int *)malloc(sizeof(int) * ((ms->nb_cmd - 1) * 2 + 2));
-	ms->pipe[0] = 0;
-	ms->pipe[(ms->nb_cmd - 1) * 2 + 1] = 1;
-	i = -1;
-	while (++i < ms->nb_cmd)
-	{
-		pipe(temp_pipe);
-		ms->pipe[i * 2 + 1] = temp_pipe[1];
-		ms->pipe[i * 2 + 2] = temp_pipe[0];
-	}
-}
-
-void	child_execution(t_ms *ms)
-{
-	t_cmd	*cmd;
-
-	cmd = cmd_lst_index(ms, ms->cmd_index);
-	pipe_redirection(ms);
-	redirection_in(cmd);
-	redirection_out(cmd);
-	execve(cmd->cmd_path, cmd->args, ms->envp);
-	printf("%s%s%s\n", ERR_FIRST, ERR_EXECVE, cmd->args[0]);
-	//exec fail 
-	
-}
-
-void	redirection_out(t_cmd *cmd)
-{
-	int	i;
-	int	new_fd_out;
-
-	i = -1;
-	new_fd_out = 0;
-	while (cmd->fd_out[++i] && new_fd_out != -1)
-	{
-		if (append[i] == 1)
-			new_fd_out = open(filename, O_WRONLY | O_CREAT | O_APPEND, 0644);
-		else
-			new_fd_out = open(filename, O_WRONLY | O_CREAT, 0644);
-		if (cmd->fd_out[i + 1] && new_fd_out != -1)
-			close(new_fd_out);
-	}
-	if (new_fd_out != -1)
-		dup2(new_fd_out, 1);
-}
-
-void	pipe_redirection(t_ms *ms)
-{
-	dup2(ms->pipe[ms->cmd_index * 2], 0);
-	dup2(ms->pipe[ms->cmd_index * 2 + 1], 1);
-}
-
-int	redirection_in(t_cmd *cmd)
-{
-	int	i;
-	int	new_fd_in;
-
-	i = -1;
-	new_fd_in = 0;
-	while (cmd->fd_in[++i] && new_fd_in != -1)
-	{
-		if (cmd->heredoc == 1)
-			new_fd_in = here_doc(cmd->fd_in[i]);
-		else
-			new_fd_in = open_fd_in(cmd->fd_in[i]);
-		if (new_fd_in != -1 && cmd->fd_in[i + 1])
-			close(new_fd_in);
-	}
-	if (new_fd_in != -1)
-		dup2(new_fd_in, 0);
-}
-
-int	open_fd_in(char *filename)
-{
-	int	new_fd_in;
-	int	error;
-
-	error = 0;
-	if (access(filename, F_OK) == -1)
-		error = openerr_nosuch;
-	else if (access(filename, X_OK) == -1)
-		error = openerr_perm;
-	else
-		new_fd_in = open(filename, O_RDONLY);
-	return (print_open_err(filename, error));
-}
-
-int	print_open_err(char *filename, int error)
-{
-	if (error = 0)
-		return (0);
-	printf("%s%s", ERR_FIRST, filename);
-	if (error = openerr_nosuch)
-		printf("%s\n", ERR_OPEN_NOSUCH);
-	else if (error == openerr_perm)
-		printf("%s\n", ERR_OPEN_PERM);
-	return (-1);
-}
-
-int	here_doc(char *str_eof)
-{
-	int		fd_pipe[2];
-	char	*line;
-
-	pipe(fd_pipe[2]);
-	readline("> ");
-	while (ft_strncmp(line, str_eof, ft_strlen(line)))
-	{
-		write(fd_pipe[1], line, ft_strlen(line));
-		readline("> ");
-	}
-	close(fd_pipe[1]);
-	return (fd_pipe[0]);
-}
-
-t_cmd	*cmd_lst_index(t_ms *ms, int cmd_index)
-{
-	int		i;
-	t_cmd	*temp;
-
-	temp = ms->cmds;
-	i = -1;
-	while (++i < cmd_index && temp->next)
-		temp = temp->next;
+	i = 0;
+	while (ms->child_id[i] && ms->child_id[i] != waitpid_return)
+		i++;
+	if (!ms->child_id[i])
+		return (-1);
+	return (i);
 }
