--- test
@@ -1,5 +1,0 @@
-Makefile
-include
-src
-temp
-test
